launch-telemetry-producer/requirements.txt
quixstreams==2.8.1
python-dotenv